Remove unneeded IIFE
