+++ src/ggrc-client/js/components/assessment/controls-toolbar/controls-toolbar.js
@@ -7,22 +7,18 @@ import '../../custom-attributes/custom-attributes-actions';
 import '../../object-state-toolbar/object-state-toolbar';
 import template from './controls-toolbar.mustache';
 
-(function (can, GGRC) {
-  'use strict';
-
-  can.Component.extend({
-    tag: 'assessment-controls-toolbar',
-    template: template,
-    viewModel: {
-      instance: null,
-      verifiers: [],
-      onStateChange: function (event) {
-        this.dispatch({
-          type: 'onStateChange',
-          state: event.state,
-          undo: event.undo,
-        });
-      },
+export default can.Component.extend({
+  tag: 'assessment-controls-toolbar',
+  template: template,
+  viewModel: {
+    instance: null,
+    verifiers: [],
+    onStateChange: function (event) {
+      this.dispatch({
+        type: 'onStateChange',
+        state: event.state,
+        undo: event.undo,
+      });
     },
-  });
-})(window.can, window.GGRC, window.CMS);
+  },
+});
