+++ src/ggrc-client/js/components/link_to_clipboard.js
@@ -5,41 +5,36 @@
 
 import Clipboard from 'clipboard';
 
-(function (can, $) {
-  'use strict';
-
-  can.Component.extend({
-    tag: 'clipboard-link',
-    scope: {
-      text: '@',
-      title: '@',
-      notify: '@',
-      isActive: false,
-      timeout: '@',
-      notifyText: 'Link has been copied to your clipboard.',
+export default can.Component.extend({
+  tag: 'clipboard-link',
+  viewModel: {
+    text: '@',
+    title: '@',
+    notify: '@',
+    isActive: false,
+    timeout: '@',
+    notifyText: 'Link has been copied to your clipboard.',
+  },
+  template: ['<a data-clipboard-text="{{text}}" {{#isActive}}class="active"{{/isActive}} href="#">',
+    '<i class="fa fa-link"></i> {{title}}',
+    '</a>'].join(''),
+  events: {
+    'a click': function (el, evnt) {
+      evnt.preventDefault();
     },
-    template: ['<a data-clipboard-text="{{text}}" {{#isActive}}class="active"{{/isActive}} href="#">',
-      '<i class="fa fa-link"></i> {{title}}',
-      '</a>'].join(''),
-    events: {
-      'a click': function (el, evnt) {
-        evnt.preventDefault();
-      },
-      'inserted': function (el, evnt) {
-        let timeout = this.scope.attr('timeout') || 10000;
-        this._clip = new Clipboard(el.find('a')[0]);
+    'inserted': function (el, evnt) {
+      let timeout = this.viewModel.attr('timeout') || 10000;
+      this._clip = new Clipboard(el.find('a')[0]);
 
-        this._clip.on('success', function () {
-          if (this.scope.attr('notify')) {
-            $('body').trigger('ajax:flash', {'success': this.scope.attr('notifyText')});
-          }
-          this.scope.attr('isActive', true);
-          setTimeout(function () {
-            this.scope.attr('isActive', false);
-          }.bind(this), timeout);
-        }.bind(this));
-      },
+      this._clip.on('success', function () {
+        if (this.viewModel.attr('notify')) {
+          $('body').trigger('ajax:flash', {'success': this.viewModel.attr('notifyText')});
+        }
+        this.viewModel.attr('isActive', true);
+        setTimeout(function () {
+          this.viewModel.attr('isActive', false);
+        }.bind(this), timeout);
+      }.bind(this));
     },
-  });
-
-})(window.can, window.can.$);
+  },
+});
