+++ src/ggrc-client/js/components/object-popover/related-assessment-popover.js
@@ -15,45 +15,41 @@ import './object-popover';
 import {convertValuesToFormFields} from '../../plugins/utils/ca-utils';
 import template from './related-assessment-popover.mustache';
 
-(function (can) {
-  'use strict';
-
-  /**
-   * Simple wrapper component to load Related to Parent Object Snapshots of Controls and Objectives
-   */
-  can.Component.extend({
-    tag: 'related-assessment-popover',
-    template: template,
-    viewModel: {
-      selectedAssessment: {},
-      popoverTitleInfo: 'Assessment Title: ',
-      define: {
-        hideTitle: {
-          type: Boolean,
-          value: false,
+/**
+ * Simple wrapper component to load Related to Parent Object Snapshots of Controls and Objectives
+ */
+export default can.Component.extend({
+  tag: 'related-assessment-popover',
+  template,
+  viewModel: {
+    selectedAssessment: {},
+    popoverTitleInfo: 'Assessment Title: ',
+    define: {
+      hideTitle: {
+        type: Boolean,
+        value: false,
+      },
+      popoverDirection: {
+        type: String,
+        value: 'right',
+      },
+      selectedAssessmentTitle: {
+        get: function () {
+          return this.attr('selectedAssessment.data.title');
         },
-        popoverDirection: {
-          type: String,
-          value: 'right',
+      },
+      selectedAssessmentLink: {
+        get: function () {
+          return this.attr('selectedAssessment.data.viewLink');
         },
-        selectedAssessmentTitle: {
-          get: function () {
-            return this.attr('selectedAssessment.data.title');
-          },
-        },
-        selectedAssessmentLink: {
-          get: function () {
-            return this.attr('selectedAssessment.data.viewLink');
-          },
-        },
-        selectedAssessmentFields: {
-          get: function () {
-            let caValues =
-              this.attr('selectedAssessment.data.custom_attribute_values');
-            return caValues ? convertValuesToFormFields(caValues) : [];
-          },
+      },
+      selectedAssessmentFields: {
+        get: function () {
+          let caValues =
+            this.attr('selectedAssessment.data.custom_attribute_values');
+          return caValues ? convertValuesToFormFields(caValues) : [];
         },
       },
     },
-  });
-})(window.can);
+  },
+});
