+++ src/ggrc-client/js/components/show-related-assessments-button/show-related-assessments-button.js
@@ -7,54 +7,50 @@ import '../related-objects/related-assessments';
 import template from './show-related-assessments-button.mustache';
 import {hasRelatedAssessments} from '../../plugins/utils/models-utils';
 
-(function (can, GGRC) {
-  'use strict';
-
-  can.Component.extend({
-    tag: 'show-related-assessments-button',
-    template: template,
-    viewModel: {
-      define: {
-        cssClasses: {
-          type: String,
-          get: function () {
-            return !this.attr('resetStyles') ?
-              'btn btn-lightBlue ' + this.attr('extraBtnCSS') : '';
-          },
-        },
-        resetStyles: {
-          type: Boolean,
-          value: false,
-        },
-        showTitle: {
-          type: Boolean,
-          value: true,
-        },
-        showIcon: {
-          type: Boolean,
-          value: false,
-        },
-        title: {
-          type: String,
-          get: function () {
-            return this.attr('text') || 'Assessments';
-          },
+export default can.Component.extend({
+  tag: 'show-related-assessments-button',
+  template,
+  viewModel: {
+    define: {
+      cssClasses: {
+        type: String,
+        get: function () {
+          return !this.attr('resetStyles') ?
+            'btn btn-lightBlue ' + this.attr('extraBtnCSS') : '';
         },
       },
-      instance: null,
-      state: {
-        open: false,
+      resetStyles: {
+        type: Boolean,
+        value: false,
       },
-      extraBtnCSS: '@',
-      text: '@',
-      modalTitle: 'Related Assessments',
-      showRelatedAssessments: function () {
-        this.attr('state.open', true);
+      showTitle: {
+        type: Boolean,
+        value: true,
       },
-      // Temporary put this logic on the level of Component itself
-      isAllowedToShow: function () {
-        return hasRelatedAssessments(this.attr('instance.type'));
+      showIcon: {
+        type: Boolean,
+        value: false,
+      },
+      title: {
+        type: String,
+        get: function () {
+          return this.attr('text') || 'Assessments';
+        },
       },
     },
-  });
-})(window.can, window.GGRC);
+    instance: null,
+    state: {
+      open: false,
+    },
+    extraBtnCSS: '@',
+    text: '@',
+    modalTitle: 'Related Assessments',
+    showRelatedAssessments: function () {
+      this.attr('state.open', true);
+    },
+    // Temporary put this logic on the level of Component itself
+    isAllowedToShow: function () {
+      return hasRelatedAssessments(this.attr('instance.type'));
+    },
+  },
+});
