+++ src/ggrc-client/js/components/workflow/workflow_deactivate.js
@@ -6,19 +6,14 @@
 
 import {getPageInstance} from '../../plugins/utils/current-page-utils';
 
-(function (GGRC, can) {
-  'use strict';
-
-  can.Component.extend({
-    tag: 'workflow-deactivate',
-    template: '<content/>',
-    events: {
-      click: function () {
-        let workflow = getPageInstance();
-        workflow.refresh().then(function (workflow) {
-          workflow.attr('recurrences', false).save();
-        });
-      },
+export default can.Component.extend({
+  tag: 'workflow-deactivate',
+  events: {
+    click: function () {
+      let workflow = getPageInstance();
+      workflow.refresh().then(function (workflow) {
+        workflow.attr('recurrences', false).save();
+      });
     },
-  });
-})(window.GGRC, window.can);
+  },
+});
